windows warning as errors, masquerade remaining unfixeable errors

+++ arangosh/Utils/ManagedDirectory.cpp
@@ -472,7 +472,7 @@ void ManagedDirectory::File::write(char const* data, size_t length) {
   }
 #endif
   if (isGzip()) {
-    gzwrite(_gzFile, data, length);
+    gzwrite(_gzFile, data, static_cast<unsigned int>(length));
   } else {
     ::rawWrite(_fd, data, length, _status, _path, _flags);
   }
@@ -494,7 +494,7 @@ ssize_t ManagedDirectory::File::read(char* buffer, size_t length) {
   }
 #endif
   if (isGzip()) {
-    bytesRead = gzread(_gzFile, buffer, length);
+    bytesRead = gzread(_gzFile, buffer, static_cast<unsigned int>(length));
   } else {
     bytesRead = ::rawRead(_fd, buffer, length, _status, _path, _flags);
   } // else
